// frontend/src/app/generated/model/v1Record.ts
/**
 * Homelab API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { V1RecordMeta } from './v1RecordMeta';
import { V1RecordStatus } from './v1RecordStatus';

export interface V1Record {
  generation?: number;
  id?: string;
  meta?: V1RecordMeta;
  resourceVersion?: number;
  status?: V1RecordStatus;
}