// frontend/src/app/generated/model/v1RecordMeta.ts
/**
 * Homelab API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

export interface V1RecordMeta {
  comments?: string;
  domainId?: string;
  enabled?: boolean;
  name?: string;
  priority?: number;
  ttl?: number;
  type?: string;
  value?: string;
}